+++ RAnalogs/RAnEn/R/verifyBrier.R
@@ -101,6 +101,9 @@ verifyBrier <- function(anen.ver, obs.ver, threshold, ensemble.func, ..., baseli
   
   require(verification)
   
+  stopifnot(dim(anen.ver) == 4)
+  stopifnot(dim(obs.ver) == 3)
+  
   if ( !identical(dim(anen.ver)[1:3], dim(obs.ver)[1:3]) ) {
     cat("Error: Observations and analogs have incompatible dimensions.\n")
     return(NULL)